move NewConfigMap to new package and add broken test
